+++ src/main/java/com/business/i4_be/domain/user/controller/UserController.java
@@ -1,8 +1,8 @@
-package com.business.i4_be.user.controller;
+package com.business.i4_be.domain.user.controller;
 
-import com.business.i4_be.user.dto.request.SignupRequest;
-import com.business.i4_be.user.dto.response.SignupResponse;
-import com.business.i4_be.user.service.UserService;
+import com.business.i4_be.domain.user.dto.request.SignupRequest;
+import com.business.i4_be.domain.user.dto.response.SignupResponse;
+import com.business.i4_be.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
