+++ src/main/java/com/business/i4_be/domain/user/dto/request/SignupRequest.java
@@ -1,4 +1,4 @@
-package com.business.i4_be.user.dto.request;
+package com.business.i4_be.domain.user.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
