+++ src/main/java/com/business/i4_be/domain/user/entity/Address.java
@@ -1,4 +1,4 @@
-package com.business.i4_be.user.entity;
+package com.business.i4_be.domain.user.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
