+++ src/main/java/com/business/i4_be/domain/user/service/UserService.java
@@ -1,9 +1,9 @@
-package com.business.i4_be.user.service;
+package com.business.i4_be.domain.user.service;
 
-import com.business.i4_be.user.dto.request.SignupRequest;
-import com.business.i4_be.user.dto.response.SignupResponse;
-import com.business.i4_be.user.entity.User;
-import com.business.i4_be.user.repository.UserRepository;
+import com.business.i4_be.domain.user.dto.request.SignupRequest;
+import com.business.i4_be.domain.user.dto.response.SignupResponse;
+import com.business.i4_be.domain.user.entity.User;
+import com.business.i4_be.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
